refactor(api): nest Builder impl and disambiguate from WorldBuilder
Fold the standalone package-private BuilderImpl source file into Builder as a
nested final class (kept mutable to preserve setName, which core/tests rely on
in place), so no *Impl file leaks into the public API package. The sealed
interface now permits the nested Builder.BuilderImpl.

Also sharpen the javadoc to distinguish Builder (a player trusted to modify a
world) from WorldBuilder (the fluent world-creation API), with a cross-reference
on each.

Serialization format (<uuid>,<name>) and the round-trip behaviour are unchanged.
